pyproject.toml
[tool.poetry]
name = "fiuber-back-fastapi"
version = "0.1.0"
description = "Fiuber backend in fastapi"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fiuber_back_fastapi"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
SQLAlchemy = "^1.4.41"
databases = "^0.6.1"
asyncpg = "^0.26.0"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
